Begin converting command line args to Cobra for better user experience.
