+++ cmd/cli.go
@@ -1,0 +1,21 @@
+package cmd
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+)
+
+func init() {
+	rootCmd.AddCommand(cliCmd)
+}
+
+var cliCmd = &cobra.Command{
+	Use:   "cli",
+	Short: "Run openbooks from the terminal in CLI mode.",
+	Long:  ``,
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("Hugo Static Site Generator v0.9 -- HEAD")
+		fmt.Printf("%v+\n", args)
+	},
+}
