+++ cmd/root.go
@@ -1,0 +1,25 @@
+package cmd
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/spf13/cobra"
+)
+
+var rootCmd = &cobra.Command{
+	Use:   "openbooks",
+	Short: "Test",
+	Long:  `A utility that makes searching and downloading eBooks easier.`,
+	// Run: func(cmd *cobra.Command, args []string) {
+	// 	// Do Stuff Here
+	// },
+}
+
+// Execute starts the root command handler.
+func Execute() {
+	if err := rootCmd.Execute(); err != nil {
+		fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
+	}
+}
