+++ go.mod
@@ -13,6 +13,7 @@ require (
 	github.com/nwaples/rardecode v1.1.0 // indirect
 	github.com/pierrec/lz4 v2.6.0+incompatible // indirect
 	github.com/rakyll/statik v0.1.7
+	github.com/spf13/cobra v1.1.1
 	github.com/spf13/viper v1.7.1 // indirect
 	github.com/xi2/xz v0.0.0-20171230120015-48954b6210f8 // indirect
 )
